fixing styling

+++ LEVEL3/part_8/tracker.py
@@ -4,38 +4,48 @@ import os
 from tqdm import tqdm
 import time
 
-
-#starting_balance = int(input("What is your current balance? ie R12 000.50: R"))
-
-
-
 class menuSelection():
-  
     def __init__(self):
-                           
-        self.category_list  = ["Savings", "Groceries","Utilites","Health","Insurance","Transportation","Clothing","Cellphone and Internet" , "Petrol/Diesel","Housing", "Main Job", "Side Hustle"]
+        self.loading_screen()
+                      
+        self.category_list  = ["Savings", "Groceries", "Utilites",
+                               "Health", "Insurance", "Transportation", "Clothing",
+                               "Cellphone and Internet", "Petrol/Diesel", "Housing", 
+                               "Main Job", "Side Hustle"]
 
 
         self.db =sqlite3.connect('budgettracker.db') #links us to the database
         self.cursor = self.db.cursor() #creates a cursor linked to the database 
 
         self.cursor.execute('''
-                    CREATE TABLE IF NOT EXISTS budgettracker(
-                    id INTEGER PRIMARY KEY,
-                    Date DATE,
-                    Description TEXT,
-                    Type TEXT,
-                    Category TEXT,
-                    Amount INT,
-                    Comments TEXT)
-                    ''') #creating database 
+            CREATE TABLE IF NOT EXISTS budgettracker(
+            id INTEGER PRIMARY KEY,
+            Date DATE,
+            Description TEXT,
+            Type TEXT,
+            Category TEXT,
+            Amount INT,
+            Comments TEXT)
+            ''') #creating database 
 
         self.db.commit()
-        
-        if os.path.exists("initialised.txt"):
-            print("EXISTS")
-        else:    
 
+        self.cursor.execute('''
+            CREATE TABLE IF NOT EXISTS categories(
+            id INTEGER PRIMARY KEY,
+            Date DATE,
+            Description TEXT,
+            Type TEXT,
+            Category TEXT,
+            Amount INT,
+            Comments TEXT)
+            ''') 
+        self.db.commit()
+    
+        if os.path.exists("initialised.txt"):
+            pass
+
+        else:    
             with open("initialised.txt","w") as file:
                 pass
             dummy_entries = [
@@ -62,33 +72,30 @@ class menuSelection():
                     ]
         
             self.cursor.executemany('''
-                                INSERT OR IGNORE INTO budgettracker(Date,Description,Type,Category,Amount,Comments)
-                                VALUES(?,?,?,?,?,?)
-                                ''', dummy_entries)
+                INSERT OR IGNORE INTO budgettracker(Date,Description,Type,Category,Amount,Comments)
+                VALUES(?,?,?,?,?,?)
+                ''', dummy_entries)
             
             self.db.commit()
             file.close()
-                        #we have to make it run once because it re-assigns them and removes the changes we need flag for dummy lis and pre-exisiting category list
-            self.category_list.sort()
-        
 
+            self.category_list.sort()
 
         self.menu_selection()
     
-
     def loading_screen(self):
         for i in tqdm(range(5)):
                     time.sleep(0.2)
 
     def clear_screen(self):
-        if os.name== "nt":
+        if os.name == "nt":
             os.system("cls")
         else:
             os.system("clear")
-         
+        
     def menu_selection(self):
-        time.sleep(5)
-   
+        time.sleep(2)
+
         menu_dictionary = {
         1: self.add_expense,
         2: self.view_expenses,
@@ -104,32 +111,31 @@ class menuSelection():
         12:self.quit_app,
         13:self.delete_account
 
-        } #dictionary holding all the functions from the menu
+        } # dictionary holding all the functions from the menu
 
-        while True:  #ensuring user choses an option thats possible 
+        while True:  # ensuring user choses an option thats possible 
 
             try:
                 menu_num_selection = int(input('''
-                                1.Add expense
-                                2.View expenses 
-                                3.View expenses by category 
-                                4.Add income
-                                5.View income
-                                6.View income by category 
-                                7.Add a category
-                                8.Set budget for a category 
-                                9.View budget for category 
-                                10.Set financial goals
-                                11.View progress towards financial goals
-                                12.Quit 
-                                13.Delete Account
-                                >> '''
-                                            ))
+                    1.Add expense
+                    2.View expenses 
+                    3.View expenses by category 
+                    4.Add income
+                    5.View income
+                    6.View income by category 
+                    7.Add a category
+                    8.Set budget for a category 
+                    9.View budget for category 
+                    10.Set financial goals
+                    11.View progress towards financial goals
+                    12.Quit 
+                    13.Delete Account
+                    >> '''
+                                ))
                 if menu_num_selection > len(menu_dictionary) or menu_num_selection < 1:
                     print("Incorrect option selection")
                 else: 
-                    break 
-
+                    break
             except ValueError: 
                 print("You have not entered a correct value")
 
@@ -137,7 +143,7 @@ class menuSelection():
 
 
     def create_table(self, rows):
-        #creating neater output. Need to create it so that it works everytime when displaying the db too 
+        # creating neater output. Need to create it so that it works everytime when displaying the db too 
         self.loading_screen()
         self.clear_screen()
    
@@ -148,15 +154,14 @@ class menuSelection():
 
 
     def create_table_with_own_headers(self,headers, rows):
-    #neater output 
+        # neater output 
         self.loading_screen()
         self.clear_screen()
-        
         table = tabulate(rows, headers= headers, tablefmt = "fancy_grid")
         print(table)
 
-    def add_expense(self):
 
+    def add_expense(self):
         while True:
 
             temp_date = input('''Please enter the date of the transaction ie yyyy-mm-dd: >>''')
@@ -180,22 +185,19 @@ class menuSelection():
                 except ValueError:
                     print("You have chosen the incorrect option")
 
-
-
             temp_amount = float(input('''How much was spent ie R120.05? >> -R'''))*-1
             
             temp_comments = input('''Please enter any comments you have about this transaction: >>''')
 
-        
-            #creating table for better visibility 
+            # creating table for better visibility 
             rows = [temp_date, temp_description,temp_type,self.category_list[category_selection-1],temp_amount,temp_comments]
             self.create_table(rows)
 
-            #validating input
+            # validating input
             correct_info = input("Please confirm if the following is correct[Y/n]:").lower()
             print(f'Your response: {correct_info}')
 
-            #validating if the details entered are correct
+            # alidating if the details entered are correct
             if correct_info != 'y' and correct_info != 'n':
                 self.create_table(rows)
                 correct_info = input("Incorrect option chosen.Please confirm if the following is correct[Y/n]:").lower().strip()
@@ -203,51 +205,40 @@ class menuSelection():
             elif correct_info == 'y':
                 print("Input Valid!")
             
-                #adding entry into database
+                # adding entry into database
                 self.cursor.execute('''
-                                    INSERT INTO budgettracker (Date, Description,Type, Category,Amount,Comments)
-                                    VALUES(?,?,?,?,?,?)''', rows)
+                    INSERT INTO budgettracker (Date, Description,Type, Category,Amount,Comments)
+                    VALUES(?,?,?,?,?,?)''', rows)
                 self.db.commit()
                 
                 break
             else: 
                 print("Incorrect input! Please re-enter... ")
-        
-
-
         self.menu_selection()
 
 
     def view_expenses(self):
        
-        
         self.cursor.execute('''
-                            SELECT * 
-                            FROM budgettracker
-                            WHERE Type = 'Expense'
-                            ORDER BY Date
-                            ''')
+            SELECT * 
+            FROM budgettracker
+            WHERE Type = 'Expense'
+            ORDER BY Date
+            ''')
         list_expenses = self.cursor.fetchall()
-        headers = ["Date", "Desciption", "Type","Category", "Amount(R)", "Comments"]
-        self.create_table_with_own_headers(headers,list_expenses)
+
+        headers = ["Date", "Desciption", "Type","Category", "Amount(R)", 
+            "Comments"]
+        
+        self.create_table_with_own_headers(headers ,list_expenses)
         time.sleep(2)
         self.menu_selection()
         
 
-
-
-
-
-
-
     def view_expenses_by_category(self):
-
-      
-
         while True:
                 for index, value in enumerate(self.category_list, start=1):
                     print(f"{index}. {value}")
-
                 try: 
                     category_selection = int(input("Please select a category to view ie 3: "))
 
@@ -259,14 +250,13 @@ class menuSelection():
                 except ValueError:
                     print("You have chosen the incorrect option")
 
-    
         print(self.category_list[category_selection-1])
         self.cursor.execute('''
-                            SELECT * 
-                            FROM budgettracker
-                            WHERE Category = ? AND Type = 'Expense'
-                            ORDER BY Date
-                            ''', (self.category_list[category_selection-1],) )
+            SELECT * 
+            FROM budgettracker
+            WHERE Category = ? AND Type = 'Expense'
+            ORDER BY Date
+            ''', (self.category_list[category_selection-1],) )
         
         list_expenses = self.cursor.fetchall()
 
@@ -277,12 +267,8 @@ class menuSelection():
         time.sleep(2)
         self.menu_selection()
         
-        
-        
-
 
     def add_income(self):
-        
         while True:
 
             temp_date = input('Please enter the date of the transaction ie yyyy-mm-dd: >>')
@@ -311,24 +297,23 @@ class menuSelection():
                  
             temp_comments = input('Please enter any comments you have about this transaction: >>')
 
-    
-            #creating table for better visibility 
+            # creating table for better visibility 
             rows = [temp_date, temp_description,temp_type, self.category_list[category_selection-1],temp_amount,temp_comments]
             self.create_table(rows)
 
-            #validating input
+            # validating input
             correct_info = input("Please confirm if the following is correct[Y/n]:").lower()
             print(f'Your response: {correct_info}')
 
-            #validating if the details entered are correct
+            # validating if the details entered are correct
             if correct_info != 'y' and correct_info != 'n':
-                self.create_table(headers,rows)
+                self.create_table(rows)
                 correct_info = input("Incorrect option chosen.Please confirm if the following is correct[Y/n]:").lower().strip()
                 print(f'Your response: {correct_info}')    
             elif correct_info == 'y':
                 print("Input Valid!")
             
-                #adding entry into database
+                 #adding entry into database
                 self.cursor.execute('''
                                     INSERT INTO budgettracker (Date, Description,Type, Category,Amount,Comments)
                                     VALUES(?,?,?,?,?,?)''', rows)
@@ -337,14 +322,10 @@ class menuSelection():
                 break
             else: 
                 print("Incorrect input! Please re-enter... ")
-        
-
-
         self.menu_selection()
 
 
     def view_income(self):
-
         self.cursor.execute('''
                             SELECT * 
                             FROM budgettracker
@@ -353,16 +334,13 @@ class menuSelection():
                             ''')
         list_income = self.cursor.fetchall()
         headers = ["Date", "Desciption", "Type","Category", "Amount(R)", "Comments"]
-
-
         self.create_table_with_own_headers(headers,list_income)
         time.sleep(2)
         self.menu_selection()
         
-   
+
     def view_income_by_category(self):
       
-        
         while True:
                 for index, value in enumerate(self.category_list, start=1):
                     print(f"{index}. {value}")
@@ -381,11 +359,11 @@ class menuSelection():
     
  
         self.cursor.execute('''
-                            SELECT * 
-                            FROM budgettracker
-                            WHERE Category = ? AND Type = 'Income'
-                            ORDER BY Date
-                            ''', (self.category_list[category_selection-1],) )
+            SELECT * 
+            FROM budgettracker
+            WHERE Category = ? AND Type = 'Income'
+            ORDER BY Date
+            ''', (self.category_list[category_selection-1],) )
         
         list_incomes = self.cursor.fetchall()
 
@@ -398,7 +376,6 @@ class menuSelection():
 
     def add_a_category(self):
  
- 
         new_category = input("Please enter a new category: ")
 
         edited_category_list = []
@@ -408,7 +385,6 @@ class menuSelection():
             i = i.lower()
 
             edited_category_list.append(i)
-
 
         if new_category.strip().lower() in edited_category_list:
             print("This category already exists")
@@ -425,55 +401,64 @@ class menuSelection():
         self.menu_selection()
 
 
-
     def set_budget_for_a_category(self):
         print("ENTERING set_budget_for_a_category")
         self.category_list.sort()
         print(self.category_list)
         
-
-
-
-
         
     def view_budget_for_category(self):
         print("ENTERING view_budget_for_category")
         pass
+
+
     def set_financial_goals(self):
         print("ENTERING set_financial_goals")
         pass
+
+
     def view_progress_towards_financial_goals(self):
         print("ENTERING view_progress_towards_financial_goals")
         pass
+
+
     def quit_app(self):
         print("Goodbye...")
         exit()
-        pass
+     
+
+
     def delete_account(self):
         confirm_delete = input("You are about to permenatley delete your account! Are you sure [Y/n]: ").lower()
 
         if confirm_delete == 'n':
-            pass
+            print("ABORTING DELETION")
+            self.menu_selection()
 
-          
+        elif confirm_delete == 'y':
 
+            print("Deleting in progress...")
 
-        self.cursor.execute('''
-                        DROP TABLE budgettracker
+            self.loading_screen()
 
-                            ''')
+            self.cursor.execute('''
+                DROP TABLE budgettracker
+                    ''')
+            
+            self.db.commit()
 
+            self.db.close()
 
-        pass
-
-
-#####################################
-mySavings = menuSelection()
-mySavings.menu_selection()
-       
-
-
-
-
-
-
+            if os.path.exists("initialised.txt"):
+                os.remove("initialised.txt")
+                print("Deleting Complete")
+                exit()
+            else: 
+                pass
+        else: 
+            print("Incorrect option selected. Returning to main menu...")
+            self.menu_selection()
+           
+# Code starts here
+new_year_resolution = menuSelection()
+new_year_resolution.menu_selection()
